+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `TensorSerializer.write_module` now accepts `include_non_persistent_buffers`
+  as a keyword-only boolean argument that can be set to `False` to exclude
+  buffers from serialization that were originally registered to the module
+  through calling `torch.nn.Module.register_buffer` with `persistent=False`
+  - `torch.nn.Module.state_dict` never includes persistent buffers,
+    so setting this to `False` will more closely match the behaviour
+    of `state_dict` serialization
+  - `TensorSerializer.write_module` used to always include non-persistent
+    buffers
+  - The default (`include_non_persistent_buffers=True`) matches the old
+    behaviour
+
 ## [2.5.1] - 2023-10-17
 
 ### Changed
@@ -190,6 +206,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `get_gpu_name`
   - `no_init_or_tensor`
 
+[Unreleased]: https://github.com/coreweave/tensorizer/compare/v2.5.1...HEAD
 [2.5.1]: https://github.com/coreweave/tensorizer/compare/v2.5.0...v2.5.1
 [2.5.0]: https://github.com/coreweave/tensorizer/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/coreweave/tensorizer/compare/v2.3.0...v2.4.0